# samples/broken-api-app/openapi.yaml
openapi: 3.0.0
info:
  title: Broken API App
  version: 1.0.0
  description: Sample API with intentional drift for testing ContractProof

servers:
  - url: http://localhost:3000/api
    description: Local development server

paths:
  /users:
    get:
      operationId: getUsers
      summary: Get all users
      tags:
        - users
      parameters:
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
    
    post:
      operationId: createUser
      summary: Create a new user
      tags:
        - users
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserInput'
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'

  /users/{id}:
    get:
      operationId: getUserById
      summary: Get user by ID
      tags:
        - users
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: User details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: User not found

  /products:
    get:
      operationId: getProducts
      summary: Get all products
      tags:
        - products
      responses:
        '200':
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'

  /orders:
    get:
      operationId: getOrders
      summary: Get all orders
      tags:
        - orders
      responses:
        '200':
          description: List of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'
    
    post:
      operationId: createOrder
      summary: Create a new order
      tags:
        - orders
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderInput'
      responses:
        '201':
          description: Order created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'

components:
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
          format: email
        createdAt:
          type: string
          format: date-time
    
    UserInput:
      type: object
      required:
        - name
        - email
      properties:
        name:
          type: string
        email:
          type: string
          format: email
    
    Product:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        price:
          type: number
        description:
          type: string
    
    Order:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        productId:
          type: string
        quantity:
          type: integer
        status:
          type: string
          enum: [pending, completed, cancelled]
        createdAt:
          type: string
          format: date-time
    
    OrderInput:
      type: object
      required:
        - userId
        - productId
        - quantity
      properties:
        userId:
          type: string
        productId:
          type: string
        quantity:
          type: integer
          minimum: 1

# Made with Bob